+++ src/core/keychain/index.test.ts
@@ -9,11 +9,14 @@ import {
   TypedMessage,
   recoverTypedSignature,
 } from '@metamask/eth-sig-util';
+import { mainnet } from 'viem/chains';
 import { beforeAll, expect, test } from 'vitest';
 
 import { delay } from '~/test/utils';
 
 // import { createTestWagmiClient } from '../wagmi/createTestWagmiClient';
+
+import { updateWagmiConfig } from '../wagmi';
 
 import { PrivateKey } from './IKeychain';
 
@@ -40,6 +43,7 @@ let privateKey = '';
 let password = '';
 
 beforeAll(async () => {
+  updateWagmiConfig([mainnet]);
   await delay(3000);
 });
