+++ src/core/raps/unlockAndSwap.test.ts
@@ -7,6 +7,7 @@ import {
   SwapType,
   getQuote,
 } from '@rainbow-me/swaps';
+import { mainnet } from 'viem/chains';
 import { beforeAll, expect, test } from 'vitest';
 
 import {
@@ -21,8 +22,8 @@ import {
 
 import { gasStore } from '../state';
 import { GasSpeed } from '../types/gas';
+import { updateWagmiConfig } from '../wagmi';
 import { getProvider } from '../wagmi/clientToProvider';
-// import { createTestWagmiClient } from '../wagmi/createTestWagmiClient';
 
 import { walletExecuteRap } from './execute';
 import { createUnlockAndSwapRap, estimateUnlockAndSwap } from './unlockAndSwap';
@@ -57,6 +58,7 @@ const SELECTED_GAS = {
 };
 
 beforeAll(async () => {
+  updateWagmiConfig([mainnet]);
   await delay(3000);
   doesntNeedUnlockQuote = await getQuote({
     chainId: 1,
